requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.39.1
django-cors-headers==3.2.1
django-embed-video==1.3.3
django-froala-editor==3.2.2
django-rest-auth==0.9.5
django-shopping-cart==0.1
django-taggit==1.3.0
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
phonenumbers==8.11.4
Pillow==7.2.0
python3-openid==3.1.0
pytz==2019.3
razorpay==1.2.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.8
django-tinymce==3.2.0
psycopg2==2.8.6